+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 8.3.0 (upcoming)
+## 8.3.0 (February 14, 2018)
 
 **New features:**